+++ packages/forge-cli/src/core/env.js
@@ -1,6 +1,7 @@
 const fs = require('fs');
 const os = require('os');
 const path = require('path');
+const inquirer = require('inquirer');
 const shell = require('shelljs');
 const { RpcClient, parseConfig } = require('@arcblock/forge-sdk');
 const debug = require('debug')(require('../../package.json').name);
@@ -17,6 +18,11 @@ const config = { cli: {} }; // global shared forge-cli run time config
 
 function setupEnv(args, requirements) {
   debug('setupEnv.args', { args, requirements });
+
+  const hasRequirements = Object.keys(requirements).some(x => requirements[x]);
+  if (hasRequirements) {
+    shell.echo(new inquirer.Separator().line);
+  }
   ensureRequiredDirs();
 
   if (requirements.forgeRelease) {
@@ -26,43 +32,43 @@ function setupEnv(args, requirements) {
   if (requirements.rpcClient) {
     ensureRpcClient(args);
   }
+
+  if (hasRequirements) {
+    shell.echo(new inquirer.Separator().line);
+  }
 }
 
+/**
+ * Ensure we have a forge release to work with, in which we find forge bin
+ *
+ * @param {*} args
+ * @param {boolean} [exitOn404=true]
+ * @returns
+ */
 function ensureForgeRelease(args, exitOn404 = true) {
-  if (args.forgeReleaseDir) {
-    if (fs.existsSync(args.forgeReleaseDir)) {
-      const forgeBinPath = path.join(args.forgeReleaseDir, './bin/forge');
-      if (fs.existsSync(forgeBinPath) && fs.statSync(forgeBinPath).isFile()) {
-        config.cli.forgeReleaseDir = args.forgeReleaseDir;
-        config.cli.forgeBinPath = forgeBinPath;
-        shell.echo(`Using forge bin path ${forgeBinPath}`);
-        return true;
-      } else {
-        shell.echo(`${symbols.error} --forge-release-dir is invalid, non forge bin found`);
-        if (exitOn404) {
-          process.exit(1);
-        }
-      }
+  const envReleaseDir = process.env.FORGE_RELEASE_DIR;
+  const cliReleaseDir = requiredDirs.release;
+  const argReleaseDir = args.releaseDir;
+
+  const releaseDir = argReleaseDir || envReleaseDir || cliReleaseDir;
+  if (fs.existsSync(releaseDir)) {
+    const forgeBinPath = path.join(releaseDir, './bin/forge');
+    if (fs.existsSync(forgeBinPath) && fs.statSync(forgeBinPath).isFile()) {
+      config.cli.releaseDir = releaseDir;
+      config.cli.forgeBinPath = forgeBinPath;
+      shell.echo(`${symbols.success} Using forge release dir: ${releaseDir}`);
+      shell.echo(`${symbols.success} Using forge executable: ${forgeBinPath}`);
+      return true;
     } else {
-      shell.echo(`${symbols.error} --forge-release-dir does not exist`);
+      shell.echo(`${symbols.error} forge release dir invalid, non forge executable found`);
       if (exitOn404) {
         process.exit(1);
       }
     }
   } else {
-    const releaseDir = requiredDirs.release;
-    const forgeBinPath = path.join(releaseDir, './bin/forge');
-    if (fs.existsSync(forgeBinPath) && fs.statSync(forgeBinPath).isFile()) {
-      config.cli.forgeReleaseDir = releaseDir;
-      config.cli.forgeBinPath = forgeBinPath;
-      shell.echo(`Using forge bin path ${forgeBinPath}`);
-      return true;
-    } else {
-      shell.echo(`${symbols.error} forge release not found under ${releaseDir}`);
-      shell.echo(`${symbols.info}, to get a forge-core release, run \`forge init\` first`);
-      if (exitOn404) {
-        process.exit(1);
-      }
+    shell.echo(`${symbols.error} forge release dir does not exist`);
+    if (exitOn404) {
+      process.exit(1);
     }
   }
 
@@ -76,19 +82,21 @@ function ensureForgeRelease(args, exitOn404 = true) {
  * @param {*} args
  */
 function ensureRpcClient(args) {
-  const configPath = args.forgeConfigPath || process.env.FORGE_CONFIG || findReleaseConfig();
+  const releaseConfig = path.join(path.dirname(requiredDirs.release), 'forge_release.toml');
+  const configPath = args.configPath || process.env.FORGE_CONFIG || releaseConfig;
   if (configPath && fs.existsSync(configPath)) {
     const forgeConfig = parseConfig(configPath);
-    shell.echo(`${symbols.success} using forge config from ${configPath}`);
+    config.cli.forgeConfigPath = configPath;
+    shell.echo(`${symbols.success} Using forge config: ${configPath}`);
     createRpcClient(forgeConfig);
-  } else if (args.forgeSocketGrpc) {
+  } else if (args.socketGrpc) {
     const forgeConfig = {
       forge: {
         decimal: 16,
-        sockGrpc: args.forgeSocketGrpc,
+        sockGrpc: args.socketGrpc,
       },
     };
-    shell.echo(`${symbols.info} using forge-cli with remote node ${args.forgeSocketGrpc}`);
+    shell.echo(`${symbols.info} using forge-cli with remote node ${args.socketGrpc}`);
     createRpcClient(forgeConfig);
   } else {
     shell.echo(`${symbols.error} forge-cli requires an forge config file to start`);
@@ -101,13 +109,12 @@ function ensureRpcClient(args) {
  *
  * @returns String
  */
-function findReleaseConfig() {
-  const { forgeReleaseDir } = config.cli;
-  if (!forgeReleaseDir) {
+function findReleaseConfig(releaseDir) {
+  if (!releaseDir) {
     return '';
   }
 
-  const libDir = path.join(forgeReleaseDir, 'lib');
+  const libDir = path.join(releaseDir, 'lib');
   if (!fs.existsSync(libDir) || !fs.statSync(libDir).isDirectory()) {
     return '';
   }
@@ -148,6 +155,18 @@ function createRpcClient(forgeConfig) {
   debug(`${symbols.info} rpc client created!`);
 }
 
+function runNativeForgeCommand(subCommand) {
+  return function() {
+    const { forgeBinPath, forgeConfigPath } = config.cli;
+    if (!forgeBinPath) {
+      shell.echo(`${symbols.error} forgeBinPath not found, abort!`);
+      return;
+    }
+
+    return shell.exec(`FORGE_CONFIG=${forgeConfigPath} ${forgeBinPath} ${subCommand}`);
+  };
+}
+
 function writeCache(key, data) {
   try {
     fs.writeFileSync(path.join(requiredDirs.cache, `${key}.json`), JSON.stringify(data));
@@ -179,7 +198,9 @@ module.exports = {
 
   setupEnv,
   requiredDirs,
+  findReleaseConfig,
   ensureRequiredDirs,
   ensureForgeRelease,
   ensureRpcClient,
+  runNativeForgeCommand,
 };